pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roomba-player"
version = "0.2.0"
description = "Roomba control, monitoring and autopilot platform"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "jvexiau" }
]
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "websockets>=12.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "python-dotenv>=1.0.1",
  "PyYAML>=6.0.1",
  "pyserial>=3.5"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"
